Refactored components to allow styleguide with them. Also moved the router-link logic to app and not the component
