+++ styleguide.config.js
@@ -1,5 +1,10 @@
+const path = require('path')
+
 module.exports = {
   components: 'src/components/**/*.vue',
-  defaultExample: true,
-  webpackConfig: require('./build/webpack.dev.conf')
+  skipComponentsWithoutExample: true,
+  webpackConfig: require('./build/webpack.dev.conf'),
+  require: [
+    path.join(__dirname, 'src/assets/common.scss')
+  ]
 }
